admin-update

+++ src/admin/content/applications/AuthorManagement/RecentAuthorsTable.tsx
@@ -32,9 +32,19 @@ import WarningIcon from '@mui/icons-material/Warning';
 interface RecentAuthorsTableProps {
   className?: string;
   authors: Author[];
+  totalPages: number;
+  totalElements: number;
+  onPageChange: (newPage: number) => void;
+  onLimitChange: (newLimit: number) => void;
 }
 
-const RecentAuthorsTable: FC<RecentAuthorsTableProps> = ({ authors }) => {
+const RecentAuthorsTable: FC<RecentAuthorsTableProps> = ({
+  authors,
+  totalPages,
+  totalElements,
+  onPageChange,
+  onLimitChange,
+}) => {
   const navigate = useNavigate();
   const cookies = new Cookies();
   const token = cookies.get('accessToken');
@@ -59,36 +69,42 @@ const RecentAuthorsTable: FC<RecentAuthorsTableProps> = ({ authors }) => {
 
   const handlePageChange = (event: any, newPage: number): void => {
     setPage(newPage);
+    onPageChange(newPage);
   };
 
   const handleLimitChange = (event: ChangeEvent<HTMLInputElement>): void => {
-    setLimit(parseInt(event.target.value));
+    const newLimit = parseInt(event.target.value);
+    setLimit(newLimit);
+    onLimitChange(newLimit);
   };
   const handleConfirmAuthor = (authorId: number) => {
-    if (token) {
-      axios
-        .post(
-          API_URL + `authors/${authorId}/confirm`,
-          {},
-          {
-            headers: {
-              Authorization: `Bearer ${token}`,
-              'Content-Type': 'application/json',
+    const confirm = window.confirm('Bạn có chắc chắn muốn xác nhận cho người dùng này không?');
+    if (confirm) {
+      if (token) {
+        axios
+          .post(
+            API_URL + `authors/${authorId}/confirm`,
+            {},
+            {
+              headers: {
+                Authorization: `Bearer ${token}`,
+                'Content-Type': 'application/json',
+              },
             },
-          },
-        )
-        .then((res) => {
-          handleShowSuccess('Thành công');
-          dispatch(deleteAny(true));
-          setTimeout(() => {
-            dispatch(resetDeleteStatus());
-          }, 2000);
-        })
-        .catch((err) => {
-          handleShowError(err.message);
-        });
-    } else {
-      handleShowError('Vui lòng đăng nhập!');
+          )
+          .then((res) => {
+            handleShowSuccess('Thành công');
+            dispatch(deleteAny(true));
+            setTimeout(() => {
+              dispatch(resetDeleteStatus());
+            }, 2000);
+          })
+          .catch((err) => {
+            handleShowError(err.message);
+          });
+      } else {
+        handleShowError('Vui lòng đăng nhập!');
+      }
     }
   };
   const handleDeleteAuthor = (authorId: number) => {
@@ -253,7 +269,7 @@ const RecentAuthorsTable: FC<RecentAuthorsTableProps> = ({ authors }) => {
       <Box p={2}>
         <TablePagination
           component="div"
-          count={authors.length}
+          count={totalElements}
           onPageChange={handlePageChange}
           onRowsPerPageChange={handleLimitChange}
           page={page}
